requirements.txt
# ============================================================================
# CORE FRAMEWORK
# ============================================================================
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-dotenv==1.0.0
gunicorn==21.2.0

# ============================================================================
# DATABASE - psycopg2 ONLY (Python 3.11 compatible)
# ============================================================================
psycopg2-binary==2.9.9

# ============================================================================
# WEBSOCKETS (threading mode, no eventlet)
# ============================================================================
python-socketio==5.10.0

# ============================================================================
# AUTHENTICATION & SECURITY
# ============================================================================
PyJWT==2.8.0
bcrypt==4.1.1
cryptography==41.0.7

# ============================================================================
# HTTP & UTILITIES
# ============================================================================
requests==2.31.0
python-dateutil==2.8.2

# ============================================================================
# IMAGE PROCESSING
# ============================================================================
Pillow==10.1.0
pytz==2024.1


